requirements.txt
fastapi
uvicorn
motor
pydantic
cloudscraper
beautifulsoup4
fake-useragent
requests
python-dotenv